+++ crc/services/lookup_service.py
@@ -18,6 +18,7 @@ from crc.models.ldap import LdapSchema
 from crc.models.workflow import WorkflowModel, WorkflowSpecDependencyFile
 from crc.services.file_service import FileService
 from crc.services.spec_file_service import SpecFileService
+from crc.services.reference_file_service import ReferenceFileService
 from crc.services.ldap_service import LdapService
 from crc.services.workflow_processor import WorkflowProcessor
 
@@ -52,7 +53,7 @@ class LookupService(object):
 
     @staticmethod
     def get_lookup_model_for_file_data(file_id, file_name, value_column, label_column):
-        file_data = SpecFileService.get_reference_file_data(file_name)
+        file_data = ReferenceFileService.get_reference_file_data(file_name)
         lookup_model = db.session.query(LookupFileModel).filter(LookupFileModel.file_model_id == file_id).first()
         if not lookup_model:
             logging.warning("!!!! Making a very expensive call to update the lookup model.")
